Footer logos, quote selection and update, images
Added in the logos for the schools and departments in the footer. Added
images into the information page (namely books).

The quote on the first page now updates/refreshes when the placeholder
yolk is clicked, and the quotes are taken from different arrays
according to the user's initial mood selection.