Starting responsiveness
I messed up with using the grid (started from the design stage - I used 10 of the 12 grid and that was a mistake). This meant I had to figure out how to deal with the responsiveness.

I have a really hacky/janky way of doing it - essentially, the offset-by-one's that I had to make up for the forced margins would have to go, and so I'm targeting specific div elements by ID to change their spacing and widths. I should've thought this through but this is the next best thing I could think of without going through and re-coding the website based on a 12-grid.

Found out that jQuery has a $.when().done(); function which is SUPER AWESOME AND USEFUL to avoid FOUC.